+++ packages/fpdart/example/src/map/overview.dart
@@ -1,5 +1,4 @@
-import 'package:fpdart/src/date.dart';
-import 'package:fpdart/src/extension/map_extension.dart';
+import 'package:fpdart/fpdart.dart';
 
 void main() {
   final d1 = DateTime(2001, 1, 1);
@@ -9,8 +8,14 @@ void main() {
   ///
   /// The first date `d1` will be overwritten by the second date `d2`,
   /// since the year is the same.
-  final map = <DateTime, int>{}.upsertAt(dateEqYear, d1, 1).upsertAt(
-        dateEqYear,
+  final map = <DateTime, int>{}
+      .upsertAt(
+        Eq.dateEqYear,
+        d1,
+        1,
+      )
+      .upsertAt(
+        Eq.dateEqYear,
         d2,
         2,
       );